List books refactor
