+++ app.rb
@@ -1,6 +1,7 @@
 require_relative 'book'
 require_relative 'student'
 require_relative 'teacher'
+require_relative 'list_book'
 
 class App
   attr_reader :people
@@ -128,17 +129,6 @@ class App
     rental = Rental.new(date, @books[selected_id_book], @people[selected_id_person])
     @rentals.push(rental)
     puts 'Rental is created successfully'
-    run
-  end
-
-  def list_books
-    puts ''
-    puts 'List of all the books in the library:'
-    puts ''
-    @books.each do |book|
-      puts "Title: #{book.title}"
-      puts "Author: #{book.author}"
-    end
     run
   end
